+++ omod/src/main/java/org/bahmni/module/hip/web/controller/BundledMedicationRequestController.java
@@ -1,13 +1,11 @@
 package org.bahmni.module.hip.web.controller;
 
 import org.bahmni.module.hip.web.exception.RequestParameterMissingException;
-import org.bahmni.module.hip.web.model.ErrorResponse;
 import org.bahmni.module.hip.web.service.BundleMedicationRequestService;
 import org.hl7.fhir.r4.model.Bundle;
 import org.openmrs.module.webservices.rest.web.RestConstants;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpHeaders;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
@@ -19,7 +17,6 @@ import static org.bahmni.module.hip.web.model.serializers.BundleSerializer.seria
 @RestController
 @RequestMapping(value = "/rest/" + RestConstants.VERSION_1 + "/hip")
 public class BundledMedicationRequestController {
-
     private BundleMedicationRequestService bundledMedicationRequestService;
 
     @Autowired
@@ -42,27 +39,6 @@ public class BundledMedicationRequestController {
         return ResponseEntity.ok()
                 .header(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_JSON_VALUE)
                 .body(serializeBundle(bundle));
-    }
-
-    @ResponseStatus(HttpStatus.BAD_REQUEST)
-    @ExceptionHandler(RequestParameterMissingException.class)
-    public @ResponseBody
-    ErrorResponse missingRequestParameter(Exception ex) {
-        return new ErrorResponse(ex.getMessage());
-    }
-
-    @ResponseBody
-    @ResponseStatus(HttpStatus.BAD_REQUEST)
-    @ExceptionHandler(IllegalArgumentException.class)
-    public String illegalArgumentException(Exception ex) {
-        return ex.getMessage();
-    }
-
-    @ResponseBody
-    @ResponseStatus(HttpStatus.BAD_REQUEST)
-    @ExceptionHandler(Exception.class)
-    public String genericException(Exception ex) {
-        return "Something went wrong!!";
     }
 
 }
