+++ omod/src/main/java/org/bahmni/module/hip/web/controller/HipControllerAdvice.java
@@ -1,0 +1,37 @@
+package org.bahmni.module.hip.web.controller;
+
+import org.apache.log4j.Logger;
+import org.bahmni.module.hip.web.exception.RequestParameterMissingException;
+import org.bahmni.module.hip.web.model.ErrorResponse;
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ControllerAdvice;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.web.bind.annotation.ResponseStatus;
+
+@ControllerAdvice
+public class HipControllerAdvice {
+    private static final Logger log = Logger.getLogger(HipControllerAdvice.class);
+
+    @ResponseStatus(HttpStatus.BAD_REQUEST)
+    @ExceptionHandler(RequestParameterMissingException.class)
+    public @ResponseBody
+    ErrorResponse missingRequestParameter(Exception ex) {
+        return new ErrorResponse(ex.getMessage());
+    }
+
+    @ResponseBody
+    @ResponseStatus(HttpStatus.BAD_REQUEST)
+    @ExceptionHandler(IllegalArgumentException.class)
+    public String illegalArgumentException(Exception ex) {
+        return ex.getMessage();
+    }
+
+    @ResponseBody
+    @ResponseStatus(HttpStatus.BAD_REQUEST)
+    @ExceptionHandler(Exception.class)
+    public ErrorResponse genericException(Exception ex) {
+        log.error(ex.getMessage(), ex);
+        return new ErrorResponse(ex.getMessage());
+    }
+}
